+++ lifoo/lifoo/src/main/java/ga/lifoo/src/comment/models/CommentList.java
@@ -4,17 +4,18 @@ import lombok.Data;
 
 import java.math.BigInteger;
 
+
 @Data
 public class CommentList {
 
-    private BigInteger commentIdx;
-    private BigInteger commentUserIdx;
+    private Integer commentIdx;
+    private Integer commentUserIdx;
     private String commentNickname;
     private String commentBody;
     private BigInteger likeNum;
     private String isLikeClicked;
 
-    public CommentList(BigInteger commentIdx, BigInteger commentUserIdx, String commentNickname, String commentBody, BigInteger likeNum, String isLikeClicked) {
+    public CommentList(Integer commentIdx, Integer commentUserIdx, String commentNickname, String commentBody, BigInteger likeNum, String isLikeClicked) {
         this.commentIdx = commentIdx;
         this.commentUserIdx = commentUserIdx;
         this.commentNickname = commentNickname;
